backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# Langchain & AI
langchain==0.3.7
langchain-openai==0.2.8
langchain-pinecone==0.2.0
langchain-community==0.3.7

# Azure OpenAI
openai>=1.52.0,<2.0.0

# Vector Store
pinecone-client==5.0.1

# Text-to-Speech
gTTS==2.5.3

# Utilities
python-dotenv==1.0.1
requests==2.32.3
aiofiles==24.1.0

# CORS
fastapi-cors==0.0.6

